+++ compose.yml
@@ -39,7 +39,6 @@ services:
       - node_modules_ghost_custom-fonts:/home/ghost/ghost/custom-fonts/node_modules:delegated
       - node_modules_ghost_custom-theme-settings-service:/home/ghost/ghost/custom-theme-settings-service/node_modules:delegated
       - node_modules_ghost_domain-events:/home/ghost/ghost/domain-events/node_modules:delegated
-      - node_modules_ghost_donations:/home/ghost/ghost/donations/node_modules:delegated
       - node_modules_ghost_email-addresses:/home/ghost/ghost/email-addresses/node_modules:delegated
       - node_modules_ghost_email-service:/home/ghost/ghost/email-service/node_modules:delegated
       - node_modules_ghost_email-suppression-list:/home/ghost/ghost/email-suppression-list/node_modules:delegated
@@ -166,7 +165,6 @@ volumes:
   node_modules_ghost_custom-fonts: {}
   node_modules_ghost_custom-theme-settings-service: {}
   node_modules_ghost_domain-events: {}
-  node_modules_ghost_donations: {}
   node_modules_ghost_email-addresses: {}
   node_modules_ghost_email-service: {}
   node_modules_ghost_email-suppression-list: {}
