+++ ghost/core/core/server/services/staff/StaffServiceEmails.js
@@ -263,7 +263,7 @@ class StaffServiceEmails {
     /**
      *
      * @param {object} eventData
-     * @param {import('@tryghost/donations').DonationPaymentEvent} eventData.donationPaymentEvent
+     * @param {import('../donations/service/DonationPaymentEvent').DonationPaymentEvent} eventData.donationPaymentEvent
      *
      * @returns {Promise<void>}
      */
