--- ghost/donations/package.json
@@ -1,32 +1,0 @@
-{
-  "name": "@tryghost/donations",
-  "version": "0.0.0",
-  "repository": "https://github.com/TryGhost/Ghost/tree/main/packages/donations",
-  "author": "Ghost Foundation",
-  "private": true,
-  "main": "build/index.js",
-  "types": "build/index.d.ts",
-  "scripts": {
-    "dev": "tsc --watch --preserveWatchOutput --sourceMap",
-    "build": "yarn build:ts",
-    "build:ts": "tsc",
-    "prepare": "tsc",
-    "test:unit": "NODE_ENV=testing c8 --src src --all --reporter text --reporter cobertura -- mocha --reporter dot -r ts-node/register './test/**/*.test.ts'",
-    "test": "yarn test:types && yarn test:unit",
-    "test:types": "tsc --noEmit",
-    "lint:code": "eslint src/ --ext .ts --cache",
-    "lint": "yarn lint:code && yarn lint:test",
-    "lint:test": "eslint -c test/.eslintrc.js test/ --ext .ts --cache"
-  },
-  "files": [
-    "build"
-  ],
-  "devDependencies": {
-    "c8": "8.0.1",
-    "mocha": "10.8.2",
-    "sinon": "15.2.0",
-    "ts-node": "10.9.2",
-    "typescript": "5.8.2"
-  },
-  "dependencies": {}
-}
