--- ghost/donations/test/.eslintrc.js
@@ -1,7 +1,0 @@
-module.exports = {
-    parser: '@typescript-eslint/parser',
-    plugins: ['ghost'],
-    extends: [
-        'plugin:ghost/test'
-    ]
-};
